Fixed compile, fixed miranda icons
git-svn-id: file:///.data/svn_convert/svn@776 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ trunk/sim/ui/toolsetup.cpp
@@ -226,7 +226,7 @@ void ToolBarSetup::removeClick()
     (*active)--;
     unsigned n = *active;
     unsigned long *p = active + (i + 1);
-    if (i < n)
+    if (i < (int)n)
         memmove(p, p + 1, (n - i) * sizeof(unsigned long));
     lstActive->removeItem(i);
     setButtons();
